# Conflicts:
#	src/API/Controller.java

+++ src/Account/AccountService.java
@@ -1,12 +1,9 @@
 package Account;
 
-import API.Controller;
-import DataBase.DataBase;
 import Money.Money;
 import Money.MoneyService;
 import Money.CurrencyDAO;
 import Person.Customer.Customer;
-import Person.Customer.CustomerService;
 import Transact.Transaction;
 import Transact.TransactionService;
 import Transact.TransactionStatus;
@@ -18,10 +15,9 @@ public class AccountService {
     private static AccountDAO accountDAO = new AccountDAO();
 
     public static boolean openAccount(Customer customer, Account account) {
-        // This function is only used to create a Checking or Saving account
+        // This function is only used to create a Checking or Saving or Security account
         // For Loan account, use the function in LoanService
-        // For Security account, use the function in SecurityService
-        if (!account.getType().equals(AccountType.CHECKING) && !account.getType().equals(AccountType.SAVING)) {
+        if (account.getType().equals(AccountType.LOAN)) {
             IO.displayMessage("AccountService: openAccount: Invalid account type", Utils.MessageType.ERROR);
             return false;
         }
@@ -31,13 +27,28 @@ public class AccountService {
         }
         account.setUsername(customer.getUsername());
         accountDAO.create(account);
+        // deposit and charge the service fee
+        Money serviceFee = new Money(Config.DEFAULT_OPEN_SERVICE_FEE, new CurrencyDAO().read("USD"));
+        deposit(account, serviceFee);
+        Money serviceFeeCharge = new Money(Config.DEFAULT_OPEN_SERVICE_FEE, new CurrencyDAO().read("USD"));
+        transfer(account, AccountService.getAccountByAccountNumber(Config.BANK_ACCOUNT_NUMBER), serviceFeeCharge);
         customer.getAccounts().add(account);
         return true;
     }
 
-    public static void closeAccount(Customer customer, Account account) {
-        customer.getAccounts().remove(account);
-        accountDAO.delete(account);
+    public static boolean closeAccount(Customer customer, Account account) {
+        // check if the account has enough money to pay the fee
+        if (account.getBalanceByCurrency("USD") >= Config.DEFAULT_CLOSE_SERVICE_FEE) {
+            // all the money in the account will be transferred to the bank
+            for (Money money : account.getCurrentBalance()) {
+                Account bankAccount = getAccountByAccountNumber(Config.BANK_ACCOUNT_NUMBER);
+                AccountService.transfer(account, bankAccount, money);
+            }
+            customer.getAccounts().remove(account);
+            accountDAO.delete(account);
+            return true;
+        }
+        return false;
     }
 
     public static boolean deposit(Account account, Money money) {
@@ -94,7 +105,13 @@ public class AccountService {
 
     public static boolean transfer(Account fromAccount, Account toAccount, Money money) {
         // start transaction
-        Transaction transaction = TransactionService.create(fromAccount, toAccount, money, TransactionType.TRANSFER);
+        Transaction transaction;
+        if (toAccount.getAccountNumber() == Config.BANK_ACCOUNT_NUMBER) {
+            IO.displayMessage("AccountService: transfer: Transfer to bank", Utils.MessageType.INFO);
+            transaction = TransactionService.create(fromAccount, toAccount, money, TransactionType.SERVICE_FEE);
+        } else {
+            transaction = TransactionService.create(fromAccount, toAccount, money, TransactionType.TRANSFER);
+        }
         // if any of the accounts is a loan account
         if (fromAccount.getType() == AccountType.LOAN || toAccount.getType() == AccountType.LOAN) {
             TransactionService.setTransactionStatus(transaction, TransactionStatus.FAILED);
@@ -103,7 +120,14 @@ public class AccountService {
         // check money list in account
         for (Money m : fromAccount.getCurrentBalance()) {
             if (m.getCurrency().getCurrencyName().equals(money.getCurrency().getCurrencyName())) {
-                if (MoneyService.update(m, -money.getAmount())) {
+                double rate = toAccount.getAccountNumber() == Config.BANK_ACCOUNT_NUMBER ? 1 : 1 + Config.DEFAULT_TRANSACT_SERVICE_FEE_RATE;
+                if (MoneyService.update(m, -money.getAmount() * rate)) {
+                    // if rate > 1, which means the bank will charge a service fee
+                    if (toAccount.getAccountNumber() != Config.BANK_ACCOUNT_NUMBER) {
+                        Account bankAccount = getAccountByAccountNumber(Config.BANK_ACCOUNT_NUMBER);
+                        Money serviceFee = new Money(money.getAmount() * Config.DEFAULT_TRANSACT_SERVICE_FEE_RATE, money.getCurrency());
+                        transfer(fromAccount, bankAccount, serviceFee);
+                    }
                     accountDAO.update(fromAccount);
                     // check money list in account
                     for (Money m2 : toAccount.getCurrentBalance()) {
